+++ MoreFertilizers/MoreFertilizers/HarmonyPatches/DomesticatedFishFood/FishPondDayUpdateTranspiler.cs
@@ -40,7 +40,7 @@ internal static class FishPondDayUpdateTranspiler
             ILHelper helper = new(original, instructions, ModEntry.ModMonitor, gen);
             helper.FindNext(new CodeInstructionWrapper[]
             {
-                new(OpCodes.Newobj, typeof(Random).GetConstructor(new[] { typeof(int) })),
+                new(OpCodes.Newobj, typeof(Random).Constructor(new[] { typeof(int) })),
                 new(SpecialCodeInstructionCases.StLoc),
             }).Advance(1);